+++ mod/src/main/java/thaumcraft/common/blocks/BlockEssentiaReservoir.java
@@ -7,12 +7,15 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 import thaumcraft.common.Thaumcraft;
 import thaumcraft.common.config.ConfigBlocks;
 import thaumcraft.common.items.wands.ItemWandCasting;
@@ -32,6 +35,10 @@ public class BlockEssentiaReservoir extends BlockContainer {
 
     @Override
     public boolean isFullCube(IBlockState state) { return false; }
+
+    @Override
+    @SideOnly(Side.CLIENT)
+    public BlockRenderLayer getRenderLayer() { return BlockRenderLayer.TRANSLUCENT; }
 
     @Override
     public EnumBlockRenderType getRenderType(IBlockState state) { return EnumBlockRenderType.MODEL; }